Role and Permissions logics for Section controller implemented.

+++ app/Http/Controllers/Admin/SectionController.php
@@ -19,6 +19,12 @@ class SectionController extends Controller
 
     public function index(Request $request): Response
     {
+        if (!auth()->user()->hasAnyRole(['admin', 'content manager'])) {
+            return response([
+                'message' => trans('section.error.index')
+            ], ResponseAlias::HTTP_FORBIDDEN);
+        }
+
         $limit = $request->query('limit', 10);
 
         $sections = Section::with('image')->paginate($limit);
@@ -30,6 +36,12 @@ class SectionController extends Controller
 
     public function show(Section $section): Response
     {
+        if (!auth()->user()->hasAnyRole(['admin', 'content manager'])) {
+            return response([
+                'message' => trans('section.error.show')
+            ], ResponseAlias::HTTP_FORBIDDEN);
+        }
+
         $section->load(['images', 'posts']);
 
         return response([
@@ -39,6 +51,12 @@ class SectionController extends Controller
 
     public function store(SectionRequest $request): Response
     {
+        if (!auth()->user()->hasAnyRole(['admin', 'content manager'])) {
+            return response([
+                'message' => trans('section.error.store')
+            ], ResponseAlias::HTTP_FORBIDDEN);
+        }
+
         $data = $request->validated();
         $images = $data['images'];
         unset($data['images']);
@@ -53,13 +71,19 @@ class SectionController extends Controller
         $section->load('images');
 
         return response([
-            'message' => trans('section.store'),
+            'message' => trans('section.success.store'),
             'section' => $section
         ], ResponseAlias::HTTP_CREATED);
     }
 
     public function update(SectionRequest $request, Section $section): Response
     {
+        if (!auth()->user()->hasAnyRole(['admin', 'content manager'])) {
+            return response([
+                'message' => trans('section.error.update')
+            ], ResponseAlias::HTTP_FORBIDDEN);
+        }
+
         $data = $request->validated();
         $images = $data['images'];
         unset($data['images']);
@@ -74,13 +98,19 @@ class SectionController extends Controller
         $section->load('images');
 
         return response([
-            'message' => trans('section.update'),
+            'message' => trans('section.success.update'),
             'section' => $section
         ], ResponseAlias::HTTP_OK);
     }
 
     public function destroy(Section $section): Response
     {
+        if (!auth()->user()->hasAnyRole(['admin', 'content manager'])) {
+            return response([
+                'message' => trans('section.error.destroy')
+            ], ResponseAlias::HTTP_FORBIDDEN);
+        }
+
         $section->images()->each(function ($image) {
             $image->delete();
         });
@@ -89,12 +119,18 @@ class SectionController extends Controller
         $section->delete();
 
         return response([
-            'message' => trans('section.destroy')
+            'message' => trans('section.success.destroy')
         ], ResponseAlias::HTTP_OK);
     }
 
     public function types(): Response
     {
+        if (!auth()->user()->hasAnyRole(['admin', 'content manager'])) {
+            return response([
+                'message' => trans('section.error.types')
+            ], ResponseAlias::HTTP_FORBIDDEN);
+        }
+
         $types = config('section.types');
 
         return response([
@@ -104,6 +140,12 @@ class SectionController extends Controller
 
     public function deleteImages(DeleteImagesRequest $request, Section $section): Response
     {
+        if (!auth()->user()->hasAnyRole(['admin', 'content manager'])) {
+            return response([
+                'message' => trans('section.error.delete_images')
+            ], ResponseAlias::HTTP_FORBIDDEN);
+        }
+
         $data = $request->validated();
 
         $section->images()->whereIn('id', $data['image_ids'])->delete();
@@ -111,7 +153,7 @@ class SectionController extends Controller
         $section->load('images');
 
         return response([
-            'message' => trans('section.images_delete'),
+            'message' => trans('section.success.images_delete'),
             'section' => $section
         ], ResponseAlias::HTTP_OK);
     }
